Expiration Reminder
A web app build with PHP and MySQL for backend, HTML and CSS for front end, on an apache server. It has many features including but not limited to: Authentication, multiple file uploads, editing capabilities, email notifications. There is purposefully no "register" option as the app did not require that and only one admin needed login capabilities for the app.

Just download and start using with your MySQL database login credentials. Create a empty folder named "filesUploaded" inside the "actions" folder. Thats it!